requirements.txt
pyautogui>=0.9.54
opencv-python>=4.8.0
Pillow>=10.0.0
keyboard>=0.13.5
PyQt5>=5.15.0
numpy>=1.24.0
pytesseract>=0.3.10
rapidocr>=3.0
scikit-learn>=1.3.0
joblib>=1.3.0
